+++ ErsatzTV/Validators/ProgramScheduleEditViewModelValidator.cs
@@ -1,9 +1,0 @@
-﻿using ErsatzTV.ViewModels;
-using FluentValidation;
-
-namespace ErsatzTV.Validators;
-
-public class ProgramScheduleEditViewModelValidator : AbstractValidator<ProgramScheduleEditViewModel>
-{
-    public ProgramScheduleEditViewModelValidator() => RuleFor(vm => vm.Name).NotEmpty();
-}